pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neonpay"
version = "1.0.0"
description = "Modern Python library for Telegram Stars (XTR) payments integration with support for Aiogram, Pyrogram, python-telegram-bot and more"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Abbas Sultanov", email = "sultanov.abas@outlook.com" }
]
keywords = [
    "telegram", "bot", "payments", "stars", "xtr", "donations", 
    "aiogram", "pyrogram", "python-telegram-bot", "telebot",
    "telegram-payments", "cryptocurrency", "micropayments"
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Communications :: Chat",
    "Topic :: Office/Business :: Financial :: Point-Of-Sale",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: AsyncIO",
]

requires-python = ">=3.9"

dependencies = [
    "aiohttp>=3.8.0",
    "typing-extensions>=4.0.0; python_version<'3.10'",
]

[project.optional-dependencies]
aiogram = ["aiogram>=3.0.0"]
pyrogram = ["pyrogram>=2.0.106", "tgcrypto>=1.2.0"]
ptb = ["python-telegram-bot>=20.0"]
telebot = ["pyTelegramBotAPI>=4.0.0"]
all = [
    "aiogram>=3.0.0",
    "pyrogram>=2.0.106", 
    "tgcrypto>=1.2.0",
    "python-telegram-bot>=20.0",
    "pyTelegramBotAPI>=4.0.0"
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
    "sphinx>=5.0.0",
    "sphinx-rtd-theme>=1.0.0"
]

[project.urls]
Homepage = "https://github.com/Abbasxan/neonpay"
Repository = "https://github.com/Abbasxan/neonpay"
Issues = "https://github.com/Abbasxan/neonpay/issues"
Documentation = "https://github.com/Abbasxan/neonpay/tree/main/docs"
Changelog = "https://github.com/Abbasxan/neonpay/releases"
Telegram = "https://t.me/neonsahib"

[tool.setuptools.packages.find]
where = ["."]
include = ["neonpay*"]
exclude = ["tests*", "examples*", "docs*"]

[tool.setuptools.package-data]
neonpay = ["py.typed"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["neonpay"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
